backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.42
pydantic==2.11.7
python-dotenv==1.0.0
email-validator==2.1.0

# Development Dependencies
pytest==7.4.3
httpx==0.25.2
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Additional Libraries
python-multipart==0.0.6
